+++ src/main/java/com/grepp/spring/app/controller/api/StudyController.java
@@ -1,12 +1,12 @@
-package com.grepp.spring.app.controller.api.study;
+package com.grepp.spring.app.controller.api;
 
-import com.grepp.spring.app.controller.api.study.payload.ApplicationRequest;
-import com.grepp.spring.app.controller.api.study.payload.ApplicationResultRequest;
-import com.grepp.spring.app.controller.api.study.payload.CheckGoalResponse;
-import com.grepp.spring.app.controller.api.study.payload.NotificationUpdateRequest;
-import com.grepp.spring.app.controller.api.study.payload.StudyCreationRequest;
-import com.grepp.spring.app.controller.api.study.payload.StudySearchRequest;
-import com.grepp.spring.app.controller.api.study.payload.StudyUpdateRequest;
+import com.grepp.spring.app.model.study.dto.request.ApplicationRequest;
+import com.grepp.spring.app.model.study.dto.request.ApplicationResultRequest;
+import com.grepp.spring.app.model.study.dto.response.CheckGoalResponse;
+import com.grepp.spring.app.model.study.dto.request.NotificationUpdateRequest;
+import com.grepp.spring.app.model.study.dto.request.StudyCreationRequest;
+import com.grepp.spring.app.model.study.dto.request.StudySearchRequest;
+import com.grepp.spring.app.model.study.dto.request.StudyUpdateRequest;
 import com.grepp.spring.app.model.chat.service.ChatService;
 import com.grepp.spring.app.model.member.dto.response.ApplicantsResponse;
 import com.grepp.spring.app.model.member.dto.response.StudyMemberResponse;
@@ -14,11 +14,11 @@ import com.grepp.spring.app.model.study.entity.Attendance;
 import com.grepp.spring.app.model.member.service.MemberService;
 import com.grepp.spring.app.model.study.code.Category;
 import com.grepp.spring.app.model.study.code.Status;
-import com.grepp.spring.app.model.study.dto.StudyCreationResponse;
-import com.grepp.spring.app.model.study.dto.StudyInfoResponse;
-import com.grepp.spring.app.model.study.dto.StudyListResponse;
-import com.grepp.spring.app.model.study.dto.WeeklyAttendanceResponse;
-import com.grepp.spring.app.model.study.dto.WeeklyGoalStatusResponse;
+import com.grepp.spring.app.model.study.dto.response.StudyCreationResponse;
+import com.grepp.spring.app.model.study.dto.response.StudyInfoResponse;
+import com.grepp.spring.app.model.study.dto.response.StudyListResponse;
+import com.grepp.spring.app.model.study.dto.response.WeeklyAttendanceResponse;
+import com.grepp.spring.app.model.study.dto.response.WeeklyGoalStatusResponse;
 import com.grepp.spring.app.model.study.reponse.GoalsResponse;
 import com.grepp.spring.app.model.study.reponse.StudyNoticeResponse;
 import com.grepp.spring.app.model.study.service.ApplicantService;
