+++ src/main/java/com/grepp/spring/app/model/study/dto/response/CheckGoalResponse.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.controller.api.study.payload;
+package com.grepp.spring.app.model.study.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
