+++ src/main/java/com/grepp/spring/app/model/study/dto/response/WeeklyAttendanceResponse.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.model.study.dto;
+package com.grepp.spring.app.model.study.dto.response;
 
 import com.grepp.spring.app.model.member.dto.response.AttendanceResponse;
 import com.grepp.spring.app.model.study.entity.Attendance;
