+++ src/main/java/com/grepp/spring/app/model/study/service/StudyService.java
@@ -1,11 +1,11 @@
 package com.grepp.spring.app.model.study.service;
 
 import com.grepp.spring.app.model.alarm.dto.request.AlarmRequest;
-import com.grepp.spring.app.controller.api.study.payload.ApplicationResultRequest;
-import com.grepp.spring.app.controller.api.study.payload.StudyCreationRequest;
-import com.grepp.spring.app.controller.api.study.payload.StudySearchRequest;
-import com.grepp.spring.app.controller.api.study.payload.StudyUpdateRequest;
-import com.grepp.spring.app.controller.api.study.payload.WeeklyAchievementCount;
+import com.grepp.spring.app.model.study.dto.request.ApplicationResultRequest;
+import com.grepp.spring.app.model.study.dto.request.StudyCreationRequest;
+import com.grepp.spring.app.model.study.dto.request.StudySearchRequest;
+import com.grepp.spring.app.model.study.dto.request.StudyUpdateRequest;
+import com.grepp.spring.app.model.study.dto.response.WeeklyAchievementCount;
 import com.grepp.spring.app.model.alarm.code.AlarmType;
 import com.grepp.spring.app.model.alarm.service.AlarmService;
 import com.grepp.spring.app.model.auth.code.Role;
@@ -21,10 +21,10 @@ import com.grepp.spring.app.model.study.code.DayOfWeek;
 import com.grepp.spring.app.model.study.code.GoalType;
 import com.grepp.spring.app.model.study.code.Status;
 import com.grepp.spring.app.model.study.code.StudyType;
-import com.grepp.spring.app.model.study.dto.StudyCreationResponse;
-import com.grepp.spring.app.model.study.dto.StudyInfoResponse;
-import com.grepp.spring.app.model.study.dto.StudyListResponse;
-import com.grepp.spring.app.model.study.dto.WeeklyGoalStatusResponse;
+import com.grepp.spring.app.model.study.dto.response.StudyCreationResponse;
+import com.grepp.spring.app.model.study.dto.response.StudyInfoResponse;
+import com.grepp.spring.app.model.study.dto.response.StudyListResponse;
+import com.grepp.spring.app.model.study.dto.response.WeeklyGoalStatusResponse;
 import com.grepp.spring.app.model.study.entity.Applicant;
 import com.grepp.spring.app.model.study.entity.GoalAchievement;
 import com.grepp.spring.app.model.study.entity.Study;
